fix: PNCP API resilience, ano_compra SMALLINT overflow, ETL error handling
- 00_download.py: abort after 20 consecutive API errors, save failed weeks
  to _checkpoint.json, verbose retry logging with error type and URL
- 04b_pncp_itens.py: TRUNCATE on start, try/except per COPY batch with
  line-by-line retry, persistent log file, _to_int() sanitizer
- 03_schema_pncp.sql, 03b_schema_pncp_itens.sql: ano_compra SMALLINT->INTEGER
- deploy.yml: continue-on-error on Full ETL step
- TODO: session 23 log, updated status

+++ sql/03b_schema_pncp_itens.sql
@@ -25,7 +25,7 @@ CREATE TABLE pncp_item (
     dt_atualizacao           TIMESTAMP,
     -- Campos de referencia (enriquecidos no download)
     cnpj_orgao               CHAR(14),
-    ano_compra               SMALLINT,
+    ano_compra               INTEGER,
     sequencial_compra        INT,
     PRIMARY KEY (numero_controle_pncp, numero_item)
 );
